update validator + mailing

+++ HTTT-SV/src/main/java/com/hd/student/service/impl/StudyRoomServiceImpl.java
@@ -1,6 +1,7 @@
 package com.hd.student.service.impl;
 
 import com.hd.student.entity.StudyRoom;
+import com.hd.student.exception.BadRequestException;
 import com.hd.student.exception.ResourceExistException;
 import com.hd.student.exception.ResourceNotFoundException;
 import com.hd.student.payload.response.ApiResponse;
@@ -13,6 +14,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,12 +25,17 @@ public class StudyRoomServiceImpl implements StudyRoomService {
     @Autowired
     private ModelMapper modelMapper;
 
+    public boolean isNameUnique(String name, Integer id) {
+        Optional<StudyRoom> existStudyRoom = studyRoomRepository.findByStudyRoomNameIgnoreCase(name);
 
+        return existStudyRoom.map(studyRoom -> studyRoom.getId().equals(id)).orElse(true);
+
+    }
     @Override
     public StudyRoomResponse addStudyRoom(StudyRoomRequest rq) {
         StudyRoom studyRoom = modelMapper.map(rq, StudyRoom.class);
-        if(studyRoomRepository.existsByStudyRoomNameIgnoreCase(rq.getStudyRoomName()))
-            throw new ResourceExistException("StudyRoomName", studyRoom.getStudyRoomName());
+        if(isNameUnique(studyRoom.getStudyRoomName(), null))
+            throw new ResourceExistException("Phòng học đã tồn tại");
         studyRoomRepository.save(studyRoom);
 
         return modelMapper.map(studyRoom, StudyRoomResponse.class);
@@ -37,10 +44,10 @@ public class StudyRoomServiceImpl implements StudyRoomService {
     @Override
     public StudyRoomResponse updateStudyRoom(StudyRoomRequest rq, int id) {
         StudyRoom studyRoom = modelMapper.map(rq, StudyRoom.class);
-        if(studyRoomRepository.existsByStudyRoomNameIgnoreCase(rq.getStudyRoomName()))
-            throw new ResourceExistException("StudyRoomName", studyRoom.getStudyRoomName());
+        if(isNameUnique(studyRoom.getStudyRoomName(), id))
+            throw new ResourceExistException("Phòng học đã trùng tên");
         else if(studyRoomRepository.findById(id).isEmpty())
-            throw new ResourceNotFoundException("StudyRoom","id", id);
+            throw new ResourceNotFoundException("Phòng học không tồn tại");
         studyRoom.setId(id);
 
         studyRoomRepository.save(studyRoom);
@@ -51,9 +58,10 @@ public class StudyRoomServiceImpl implements StudyRoomService {
 
     @Override
     public ApiResponse deleteStudyRoom(int id) {
-        StudyRoom st = this.studyRoomRepository.findById(id).orElseThrow(()-> new ResourceNotFoundException("StudyRoom", "id", id));
-
-        studyRoomRepository.delete(st);
+        StudyRoom st = this.studyRoomRepository.findById(id).orElseThrow(()
+                -> new ResourceNotFoundException("Không tìm thấy phòng học"));
+        if(st.getScheduleInfos().isEmpty())
+            studyRoomRepository.delete(st);
         return new ApiResponse("Xóa thành công", true);
     }
